requirements.txt
absl-py==0.15.0
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
clang==5.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
coverage==6.0.2
cvxopt==1.2.7
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.8
docopt==0.6.2
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.41.0
gTTS==2.2.2
h5py==3.1.0
idna==2.10
image==1.5.33
iniconfig==1.1.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipythonblocks==1.9.0
ipywidgets==7.6.5
isodate==0.6.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.1.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.2
notebook==6.4.5
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
Owlready==0.3.1
Owlready2==0.33
packaging==21.0
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
playsound==1.2.2
pluggy==1.0.0
prettytable==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.19.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///E:/Download/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycparser==2.20
pygame==2.0.1
Pygments==2.10.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.18.0
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.1
pyttsx==1.1
pyttsx3==2.90
pytz==2021.1
pywin32==300
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
qpsolvers==1.7.0
qtconsole==5.1.1
QtPy==1.11.2
quadprog==0.1.10
rdflib==6.0.0
rdflib-jsonld==0.5.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
Send2Trash==1.8.0
sepy==0.43.2.1
six==1.15.0
sortedcontainers==2.4.0
SpeechRecognition==3.8.1
SQLAlchemy==1.3.24
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
tqdm==4.60.0
traitlets==5.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
yarg==0.1.9
